requirements.txt
jsonschema>=4.20.0
pyyaml>=6.0
